backend/cloud-functions/llm-query/main.py
import functions_framework
from flask import jsonify
import os
import google.cloud.aiplatform as aiplatform

# TODO: Set your Google Cloud project ID, location, and model name
PROJECT_ID = os.environ.get('GCP_PROJECT')
LOCATION = os.environ.get('FUNCTION_REGION', 'us-central1') # Default to us-central1 if not set
MODEL_NAME = "gemini-1.0-pro" # Example model, adjust as needed

@functions_framework.http
def query_llm(request):
    """HTTP Cloud Function to handle LLM queries."""
    if not PROJECT_ID:
        return jsonify({"error": "GCP_PROJECT environment variable not set."}), 500

    request_json = request.get_json(silent=True)
    if not request_json or 'query' not in request_json:
        return jsonify({"error": "Invalid request. JSON payload with 'query' key is required."}), 400

    user_query = request_json['query']

    try:
        aiplatform.init(project=PROJECT_ID, location=LOCATION)
        # TODO: Implement actual call to the Vertex AI model
        # Example placeholder for model interaction:
        # model = aiplatform.GenerativeModel(MODEL_NAME)
        # response = model.generate_content(user_query)
        # result = response.text

        # --- Mock Response for now ---
        print(f"Received query: {user_query}")
        result = f"Mock Vertex AI response for: '{user_query}'"
        sources = ["mock_source1.txt", "mock_source2.pdf"]
        # --- End Mock Response ---

        return jsonify({
            "response": result,
            "sources": sources # Placeholder for document sources
        })

    except Exception as e:
        print(f"Error processing query: {e}")
        return jsonify({"error": f"Internal server error: {e}"}), 500 