backend/cloud-functions/llm-query/requirements.txt
# Function dependencies, for example:
functions-framework==3.*
google-cloud-aiplatform>=1.38.0 